+++ CHANGELOG.md
@@ -18,6 +18,8 @@
   of `BuiltinString`, but potentially useful for "printf-debugging" of scripts.
 - An option `txOptEmulatorParamsModification` to temporarily change protocol
   parameters for balancing and validation of a transaction
+- A function `combineModsTweak` to construct branching tweaks depending on the
+  different combinations of foci of an optic on `TxSkel`
 
 ### Removed